requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.2.0
pinecone>=3.0.0
sentence-transformers>=2.2.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
# Google Gemini LLM
langchain-google-genai>=1.0.0

# PDF reading
pymupdf>=1.23.0
pypdf>=3.0.0

# Word document reading
python-docx>=1.0.0